requirements_bk_3110.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
diff-match-patch==20241021
Django==5.0.9
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-import-export==4.2.0
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
httplib2==0.22.0
idna==3.10
lxml==5.3.0
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tablib==3.7.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.3
mysql-connector-python
transformers
scikit-learn
torch
openai